requirements.txt
matplotlib==3.7.1
numpy==1.24.3
paddlepaddle_gpu==2.5.0
Pillow==9.4.0
